fixed bug where wouldn't play correct portion of the audio when clicking on a segment and wouldn't show the right red bar audio cursor location -- fix was to make sure we use 48K sample mp3 files
Former-commit-id: 89b123d161443e3e57c45f58c56a3dd8a48c5455 [formerly 28bf28a9ba025057f9e52fbef649c4e2d6f4d3e2]
Former-commit-id: 54463291601cca301503ce0dff90bf4686192e75

+++ src/mitll/langtest/client/sound/SoundManagerStatic.java
@@ -36,7 +36,7 @@ public class SoundManagerStatic implements SoundManagerAPI {
     SoundManager.setPositionAndPlay(sound, position);
   }
 
-  public void playInterval(Sound sound, double start, double end) {
+  public void playInterval(Sound sound, int start, int end) {
     SoundManager.playInterval(sound, start, end);
   }
